Flatten the binary directories (Frameworks, Plugins, and Helpers).
This makes the bundle comply fully with Apple's codesigning
recommendations. It does not seem to make any actual difference in
performance.

Also note that because Gtk insists that its modules should be in a
particular subdirectory relative to $GTK_PATH this breaks finding
those modules; Gtk would have to be patched for this to work.